Lots of debugging, images not showing properly, mismath in coordinates or something.
Finally managed to get it working again, memory allocation was failing,
Another thing that was the issue is that wrong data type was storing triangle indices.

Also issues with triangles meshes not drawing properly. Everything looked out of
allignment with the model. Now moving on I will test only body and see if it draw properly.