Implemented reading of triangles, vertices and everything.
Models seems to be read fine, the problem I am encountering is that
for 'barney.mdl' file I am reading beyond the file itself. For this I need
to make a fix. Debugging using lldb showed that I want to read the data but as the reading occures
we reach the end of the file, and this is not handled properly at the moment so we end up crashing and
dont store anything that we read